Consolidate work to ServerSockEP
