+++ src/server/ServerSockEP.h
@@ -45,7 +45,8 @@ public:
 
 protected:
     virtual int addClient(std::unique_ptr<ISSClientSockEP> newClient);
-    virtual void runServer() = 0;
+    void runServer();
+    virtual void handlePfdUpdates(const std::vector<struct pollfd> &pfds, std::vector<struct pollfd> newPfds, std::vector<struct pollfd> &removePfds) = 0;
     virtual void closeSocket();
 
     // allow concrete class to create the proper type of client
