+++ src/server/UnixDgramServerSockEP.h
@@ -19,8 +19,8 @@ public:
 
 private:
 
-    void runServer() override; // meant to be the function for the receive thread
-
+    // void runServer() override; // meant to be the function for the receive thread
+    void handlePfdUpdates(const std::vector<struct pollfd> &pfds, std::vector<struct pollfd> newPfds, std::vector<struct pollfd> &removePfds) override;
     std::unique_ptr<ISSClientSockEP> createNewClient() override;
 
     struct sockaddr_un saddr_;
